requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
sentence_transformers
faiss-cpu
requests
openpyxl
httpx
pymupdf
python-docx